requirements.txt
# AI APIs
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
sqlite3-api>=2.0.0

# Google Sheets
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0

# Video processing
moviepy>=1.0.3
Pillow>=10.0.0

# Text-to-speech
edge-tts>=6.1.0

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1
aiohttp>=3.9.0
